requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
python-dotenv==1.0.0